refactor: move user editor editing helpers to executor_admin_user_editor_edit.go
